+++ lazy_sign_in/lib/lazy_sign_in.dart
@@ -1,5 +1,4 @@
 library lazy;
 
 export 'src/sign_in.dart' show SignIn;
-export 'src/sign_in_ext_dummy.dart' show SignInDummy;
-export 'src/is_signed_in.dart' show IsSignedIn;
+export 'src/sign_in_dummy.dart' show SignInDummy;